Messages moved to interface, getting offers from lecturer's api

+++ CourierAppBackend/Models/Package.cs
@@ -8,5 +8,5 @@ public class Package
     public int Height { get; set; }
     public int Width {  get; set; }
     public int Length {  get; set; }
-    public decimal Weight { get; set; }
+    public float Weight { get; set; }
 }
